move assessment and visualization configuration to their own files
