+++ src/DetailsView/components/assessment-test-view.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import { AssessmentDefaultMessageGenerator } from '../../assessments/assessment-default-message-generator';
 import { AssessmentsProvider } from '../../assessments/types/assessments-provider';
 import { AssessmentTestResult } from '../../common/assessment/assessment-test-result';
-import { VisualizationConfiguration } from '../../common/configs/visualization-configuration-factory';
+import { VisualizationConfiguration } from '../../common/configs/visualization-configuration';
 import { NamedSFC } from '../../common/react/named-sfc';
 import { AssessmentStoreData } from '../../common/types/store-data/assessment-result-data';
 import { TabStoreData } from '../../common/types/store-data/tab-store-data';
