+++ src/DetailsView/components/left-nav/left-nav-link-builder.tsx
@@ -4,7 +4,7 @@ import { map } from 'lodash';
 import * as React from 'react';
 
 import { AssessmentsProvider } from '../../../assessments/types/assessments-provider';
-import { VisualizationConfiguration } from '../../../common/configs/visualization-configuration-factory';
+import { VisualizationConfiguration } from '../../../common/configs/visualization-configuration';
 import { ManualTestStatus, ManualTestStatusData } from '../../../common/types/manual-test-status';
 import { VisualizationType } from '../../../common/types/visualization-type';
 import { DictionaryStringTo } from '../../../types/common-types';
