+++ src/assessments/assessment-builder.tsx
@@ -8,7 +8,7 @@ import { AssessmentToggleActionPayload } from '../background/actions/action-payl
 import { createInitialAssessmentTestData } from '../background/create-initial-assessment-test-data';
 import { InstanceIdentifierGenerator, UniquelyIdentifiableInstances } from '../background/instance-identifier-generator';
 import { RequirementComparer } from '../common/assessment/requirement-comparer';
-import { AssesssmentVisualizationConfiguration } from '../common/configs/visualization-configuration-factory';
+import { AssesssmentVisualizationConfiguration } from '../common/configs/assesssment-visualization-configuration';
 import { Messages } from '../common/messages';
 import { ManualTestStatus } from '../common/types/manual-test-status';
 import { FeatureFlagStoreData } from '../common/types/store-data/feature-flag-store-data';
