+++ src/assessments/types/iassessment.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import { InitialDataCreator } from '../../background/create-initial-assessment-test-data';
 import { RequirementOrdering } from '../../common/assessment/requirement';
-import { AssesssmentVisualizationConfiguration } from '../../common/configs/visualization-configuration-factory';
+import { AssesssmentVisualizationConfiguration } from '../../common/configs/assesssment-visualization-configuration';
 import { AnyExtension } from '../../common/extensibility/extension-point';
 import { AssessmentData } from '../../common/types/store-data/assessment-result-data';
 import { VisualizationType } from '../../common/types/visualization-type';
