+++ src/injected/client-view-controller.ts
@@ -4,7 +4,8 @@ import { autobind } from '@uifabric/utilities';
 import * as _ from 'lodash';
 import { BaseStore } from '../common/base-store';
 import { TestMode } from '../common/configs/test-mode';
-import { VisualizationConfiguration, VisualizationConfigurationFactory } from '../common/configs/visualization-configuration-factory';
+import { VisualizationConfiguration } from '../common/configs/visualization-configuration';
+import { VisualizationConfigurationFactory } from '../common/configs/visualization-configuration-factory';
 import { EnumHelper } from '../common/enum-helper';
 import { AssessmentStoreData } from '../common/types/store-data/assessment-result-data';
 import { FeatureFlagStoreData } from '../common/types/store-data/feature-flag-store-data';
