+++ src/popup/components/diagnostic-view-toggle.tsx
@@ -6,7 +6,8 @@ import { Spinner, SpinnerSize } from 'office-ui-fabric-react/lib/Spinner';
 import { IToggle } from 'office-ui-fabric-react/lib/Toggle';
 import * as React from 'react';
 import { VisualizationToggle } from '../../common/components/visualization-toggle';
-import { VisualizationConfiguration, VisualizationConfigurationFactory } from '../../common/configs/visualization-configuration-factory';
+import { VisualizationConfiguration } from '../../common/configs/visualization-configuration';
+import { VisualizationConfigurationFactory } from '../../common/configs/visualization-configuration-factory';
 import { KeyCodeConstants } from '../../common/constants/keycode-constants';
 import { TelemetryEventSource } from '../../common/telemetry-events';
 import { DetailsViewPivotType } from '../../common/types/details-view-pivot-type';
