+++ src/tests/unit/common/test-assessment-provider.tsx
@@ -6,7 +6,7 @@ import * as React from 'react';
 import { AssessmentsProviderImpl } from '../../../assessments/assessments-provider';
 import { Assessment } from '../../../assessments/types/iassessment';
 import { RequirementComparer } from '../../../common/assessment/requirement-comparer';
-import { AssesssmentVisualizationConfiguration } from '../../../common/configs/visualization-configuration-factory';
+import { AssesssmentVisualizationConfiguration } from '../../../common/configs/assesssment-visualization-configuration';
 import { FeatureFlags } from '../../../common/feature-flags';
 import { ManualTestStatus } from '../../../common/types/manual-test-status';
 import { AssessmentData } from '../../../common/types/store-data/assessment-result-data';
